Initialize AWS session lazily
This allows ssm-env to skip AWS session init and region lookup if there
are no SSM params present in the environment.

This also fixes a bug where ssm-env would crash on the region lookup
when running outside of an AWS env, even if there were no params to
expand and/or if `-nofail` was passed.