Ignore errors during region lookup, set directly on existing session
It should be safe to set the region config directly on the existing
session rather than construct a new one, since it isn't being accessed
concurrently by anything else at this point.

It also is simpler to ignore any errors that happen during region lookup
since we'll get a direct (and more contextualized) error anyway if the
session doesn't have a region configured.